# packages/backend/src/todo-lists/dto/create-todo-list.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsString } from 'class-validator';

export class CreateTodoListDto {
  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  name: string;
}